Subiendo cambios locales del backend a la rama 857656_DanielDonado

+++ src/main/java/com/uniminuto/clinica/entity/Usuario.java
@@ -16,43 +16,31 @@ import lombok.Data;
  */
 @Data
 @Entity
-@Table(name = "usuario")
+@Table(name="usuario")
 public class Usuario implements Serializable {
+    
     /**
      * Id serializable.
      */
     private static final long serialVersionUID = 1L;
-    /**
-     * Id.
-     */
+    
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "id")
-    private Integer id;
-    /**
-     * Username.
-     */
+    private Long id;
+    
     @Column(name = "username")
     private String username;
-    /**
-     * Password.
-     */
+    
     @Column(name = "password_hash")
     private String password;
-    /**
-     * Rol.
-     */
+    
     @Column(name = "rol")
-    private String rol;    
-    /**
-     * Fecha Creacion.
-     */
+    private String rol;
+    
     @Column(name = "fecha_creacion")
     private LocalDateTime fechaCreacion;
     
-    /**
-     * Activo.
-     */
     @Column(name = "activo")
     private boolean activo;
